Feat: InputView 구현

+++ src/main/java/step2/domain/lotto/LottoStatistics.java
@@ -8,7 +8,7 @@ public class LottoStatistics {
         this.purchasedAmount = purchasedAmount;
     }
 
-    public double getWinningRate(int winningAmount) {
+    public double getWinningRate(long winningAmount) {
         return (double) winningAmount / purchasedAmount;
     }
 }
